make small tweaks and create PhoneRechargeClient
